chore: minor example cleanups
Small correctness and consistency fixes across the compiled-binding, Rust,
and free-threading examples.

- pyminuit2 (pybind11 + nanobind, kept mirrored): missing includes, project
  name typo, CONFIGURE_DEPENDS globs, PRIVATE linkage for MODULE targets,
  drop stray #pragma once, unify to the _a argument style, raise nanobind's
  stale Python 3.8 floor, and CONFIG REQUIRED on find_package(pybind11).
- vectorclass: getters/setters use double instead of silently truncating to
  float.
- Rust PyO3: add rust-version = "1.85" for edition 2024; raise pixi maturin
  floor to >=1.14 to match the build requirement.
- free-threading: exclude **.cpp from the pybind11/nanobind wheels and trim
  the .gitignore to what these variants need.
- content: fix the compiled intro to point at the correct chapter for Rust.

+++ examples/2_02_binding/vectorclass/VectorClass.hpp
@@ -7,11 +7,11 @@ class Vector2D {
   public:
     Vector2D(double x, double y) : x(x), y(y) {}
 
-    float get_x() const { return x; }
-    float get_y() const { return y; }
+    double get_x() const { return x; }
+    double get_y() const { return y; }
 
-    void set_x(float val) { x = val; }
-    void set_y(float val) { y = val; }
+    void set_x(double val) { x = val; }
+    void set_y(double val) { y = val; }
 
     Vector2D &operator+=(const Vector2D &other) {
         x += other.x;
